+++ gonk/Cargo.toml
@@ -1,32 +1,16 @@
 [package]
 name = "gonk"
-version = "0.0.10"
+version = "0.1.0"
 edition = "2021"
 
-authors = ["Bay"]
-description = "A terminal music player"
-repository = "https://github.com/zX3no/gonk"
-readme = "../README.md"
-license = "MIT"
-
 [dependencies]
-# Music
-gonk-player = {version = "0.0.10", path = "../gonk-player"}
-
-# UI
 crossterm = "0.23.2"
-tui = { version = "0.18.0", features = ["serde"] }
-unicode-width = "0.1.9"
+tui = "0.18.0"
 
-# Config
-static_init = "1.0.2"
-serde = { version = "1.0.137", features = ["derive"] }
-toml = { version = "0.5.9", features = ["preserve_order"] }
-
-# Database
 rusqlite = { version = "0.27.0", features = ["bundled"] }
 jwalk = "0.6.0"
 rayon = "1.5.3"
+static_init = "1.0.2"
 
-# Search
-strsim = "0.10.0"
+gonk-player = {path = "../gonk-player"}
+unicode-width = "0.1.9"
